% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corstars}
\alias{corstars}
\title{Correlation matrix}
\usage{
corstars(
  cor_value,
  cor_p,
  removeTriangle = c("lower"),
  result = c("none", "html", "latex")
)
}
\arguments{
\item{cor_value}{association matrix}

\item{cor_p}{significance of association test}

\item{removeTriangle}{display correlation table as upper or lower triangular matrix}

\item{result}{output format as none, html or latex}
}
\value{
Correlation coefficient and significance of correlation test between each pair of varaibles
}
\description{
Calculate pairwise correlation matrix with significance
}
\examples{
data(mtcars)
types <- rep("numeric", ncol(mtcars))
test <- pairwise_cor(mtcars, types)
corstars <-function(test$cor_value, test$cor_p)

}
