% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_r2}
\alias{get_r2}
\title{R-squared}
\usage{
get_r2(df, type)
}
\arguments{
\item{df}{data frame}

\item{type}{type corresponding to columns in df}
}
\value{
A vector with values of R squared of each variable
}
\description{
Compute R-squared for each variable versus all other covariates
}
\details{
R-squared is a measure of collinearity,
defined as how one variable is linearly associated with all other variables in the data set.
}
\examples{
df <- data.frame(X1 = rnorm(100), X2 = rnorm(100), X3 = rnorm(100))
get_r2j(df, rep("numeric", 3))
}
