% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_r2j}
\alias{get_r2j}
\title{Adjusted R-squared}
\usage{
get_r2j(df)
}
\arguments{
\item{df}{data frame or matrix}
}
\value{
A table with values of adjusted R squared of each variable
}
\description{
Compute adjusted R-squared for each variable
}
\details{
Adjusted R-squared is a measure of collinearity,
defined as how one variable is linearly associated with all other variables in the data set.
}
\examples{
df <- data.frame(X1 = rnorm(100), X2 = rnorm(100), X3 = rnorm(100))
get_r2j(df)
}
