% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot_wsig.R
\name{network_plot_wsig}
\alias{network_plot_wsig}
\title{Create a correlation network plot with significance lines}
\usage{
network_plot_wsig(
  df,
  method = "spearman",
  sig.level = 0.05,
  min_cor = 0.3,
  legend = TRUE,
  overlay = TRUE,
  colours = c("indianred2", "white", "skyblue1"),
  repel = TRUE,
  colors,
  label_size = 5
)
}
\arguments{
\item{df}{the data set to visualize}

\item{method}{spearman or pearson}

\item{min_cor}{the threshold correlation to display}

\item{legend}{legend for correaltion}

\item{colours}{see network_plot}

\item{repel}{see network_plot}

\item{colors}{see network_plot}

\item{sig, level}{the level at which to display significance lines}
}
\description{
Create a correlation network plot with significance lines
}
