% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{pairwise_cor}
\alias{pairwise_cor}
\title{Calculate pariwise association of data with mixed types of variables}
\usage{
pairwise_cor(df, var_type = NULL)
}
\arguments{
\item{df}{dataframe with mixed types of variables}

\item{var_type}{a charater vector corresponding to types of variables in df
if not provided, will guess.}
}
\value{
matrices with measures, types and p values of association
}
\description{
Calculate pariwise association of data with mixed types of variables
}
\details{
The following associated measures and tests are implemented dependent on variable type:

factor vs numeric, factor or ordinal: Pseudo R^2 and p value from multinomial regression

ordinal vs ordinal or numeric: GK gamma and GK gamma correlation test

numeric vs numeric: Spearman correlation and p value
}
\examples{
data1 <- data.frame(x = rnorm(10),
 y = rbinom(10, 1, 0.5),
z = rbinom(10, 5, 0.5))
type1 <- c("numeric", "factor", "ordinal")
pairwise_cor(data1, type1)

}
