% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{update_reactive_df}
\alias{update_reactive_df}
\title{Update the data reactively}
\usage{
update_reactive_df(df_lst)
}
\arguments{
\item{df_lst}{}
}
\value{

}
\description{
Update the data reactively
}
