% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{update_transform_df}
\alias{update_transform_df}
\title{Update data after pre-processing operation}
\usage{
update_transform_df(df_lst)
}
\arguments{
\item{df_lst}{A list of datasets}
}
\value{
A list of pre-processed datasets
}
\description{
Update data sets reactively with pre-processing operations specified by user,
as well as input panel in following tabs
}
\details{
This function only run internally in VisX application because it depends on
 reactive values from user input
}
