% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corstars}
\alias{corstars}
\title{Create a table for correlation with significance}
\usage{
corstars(
  x,
  method = c("spearman", "pearson"),
  removeTriangle = c("lower"),
  result = c("none", "html", "latex")
)
}
\arguments{
\item{x}{a data frame or matrix}

\item{method}{spearman or pearson}

\item{removeTriangle}{display correlation table as upper or lower trianglur matrix}

\item{result}{output as none, html or latex}
}
\value{
correlation and significance of input data set
}
\description{
Create a table for correlation with significance
}
