% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corstars}
\alias{corstars}
\title{Correlation matrix}
\usage{
corstars(cor_value, cor_p, var_type)
}
\arguments{
\item{cor_value}{association matrix}

\item{cor_p}{significance of association test}

\item{var_type}{type of variables that correlation and association is calculated for}

\item{result}{output format as none, html or latex}
}
\value{
Correlation coefficient and significance of correlation test between each pair of varaibles
}
\description{
Calculate pairwise correlation matrix with significance marked by stars.
Columns and rows are grouped by variable type
}
\examples{
data(mtcars)
types <- rep("numeric", ncol(mtcars))
test <- pairwise_cor(mtcars, types)
type <- c("numeric", "factor",  rep("numeric", 5), rep("factor", 2), rep("ordinal", 2))
corstars <-function(test$cor_value, test$cor_p, type)

}
