% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corstars}
\alias{corstars}
\title{Correlation matrix}
\usage{
corstars(
  x,
  method = c("spearman", "pearson"),
  removeTriangle = c("lower"),
  result = c("none", "html", "latex")
)
}
\arguments{
\item{x}{data frame or matrix}

\item{method}{method to calculate correlation coefficient; spearman or pearson}

\item{removeTriangle}{display correlation table as upper or lower triangular matrix}

\item{result}{output format as none, html or latex}
}
\value{
Correlation coefficient and significance of correlation test between each pair of varaibles
}
\description{
Calculate pairwise correlation matrix with significance
}
\examples{
data(mtcars)
corstars(mtcars, method = "spearman")
}
