% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{corstars}
\alias{corstars}
\title{Correlation and association matrix}
\usage{
corstars(cor_value, cor_p, var_type)
}
\arguments{
\item{cor_value}{association matrix}

\item{cor_p}{significance of association test}

\item{var_type}{type of variables that correlation and association is calculated for}
}
\value{
Correlation coefficient and significance of correlation test between each pair of variables,
as well as variable type corresponding to rows and columns for display in shiny app
}
\description{
Calculate pairwise correlation matrix with significance marked by stars.
Columns and rows are grouped by variable type
}
\examples{
data(mtcars)
library(VisX)
types <- rep("numeric", ncol(mtcars))
test <- VisX::pairwise_cor(mtcars, types)
type <- c("numeric", "factor",  rep("numeric", 5), rep("factor", 2), rep("ordinal", 2))
corstars(test$cor_value, test$cor_p, type)

}
