% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{data_check}
\alias{data_check}
\title{Check for columns that are empty/identical}
\usage{
data_check(df)
}
\arguments{
\item{df}{dataframe to check}
}
\value{
a warning message if there exists empty columns or columns with no variation in the data
}
\description{
Check for columns that are empty/identical
}
