% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions.R
\name{df_for_figure}
\alias{df_for_figure}
\title{Prepare data for correlation diagram}
\usage{
df_for_figure(df_lst)
}
\arguments{
\item{df_lst}{}
}
\value{

}
\description{
Prepare data for correlation diagram
}
