% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_r2j}
\alias{get_r2j}
\title{Computed adjusted R-squared}
\usage{
get_r2j(df)
}
\arguments{
\item{df}{a data frame or matrix}
}
\value{
a table with adjusted R squared of each variables verses all other variables
}
\description{
Computed adjusted R-squared
}
