% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot_wsig.R
\name{network_plot_wsig}
\alias{network_plot_wsig}
\title{Correlation diagram}
\usage{
network_plot_wsig(
  df,
  method = "spearman",
  sig.level = 0.05,
  min_cor = 0.3,
  legend = TRUE,
  overlay = TRUE,
  colours = c("indianred2", "white", "skyblue1"),
  repel = TRUE,
  colors,
  label_size = 5
)
}
\arguments{
\item{df}{Data set to visualize}

\item{method}{Method to compute correlation coefficient; spearman or pearson}

\item{sig.level}{Significance threshold to display}

\item{min_cor}{Correlation threshold to display}

\item{legend}{Include legend in the diagram; TRUE or FALSE}

\item{colours}{Vector of colors to use for n-color gradient.}

\item{repel}{Should variable labels repel each other; TRUE or FALSE}

\item{colors}{same as colours}
}
\value{
A correlation diagram visualizing pairwise correlation between variables with network
Thickness and color hue of edges indicate strength of correlation, and
significant relationships are marked out as solid line.
}
\description{
Create a network plot for pairwise correlation between variables
}
\examples{
data(mtcars)
network_plot_wsig(mtcars, method = "spearman", sig.level = 0.05, min_cor = 0.3)
}
