% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_mixed_cor.R
\name{npc_mixed_cor}
\alias{npc_mixed_cor}
\title{Network Plot of Association}
\usage{
npc_mixed_cor(
  cor_value,
  cor_type,
  cor_p,
  var_type,
  sig.level = 0.05,
  min_cor = 0.3,
  show_signif = FALSE,
  legend = TRUE,
  repel = TRUE,
  label_size = 5,
  overlay = TRUE
)
}
\arguments{
\item{cor_value}{a matrix with values of association measures}

\item{cor_type}{a matrix with types of association measures}

\item{cor_p}{a matrix with p values of association tests}

\item{var_type}{a character vector with types of variables association is measured on}

\item{sig.level}{maximum significance to be visualized}

\item{min_cor}{minimum association to be visualized}

\item{show_signif}{whether or not to visualize significance of association}

\item{legend}{if TURE, how legend}

\item{repel}{if TRUE, text labels of variable names repel each other}

\item{label_size}{size of text labels of variable names}

\item{overlay}{if TRUE, only show significant association above min_cor threshold}
}
\value{
A network plot of association
}
\description{
Network Plot of Association
}
\examples{
data("mtcars")
type1 <- c(rep("numeric", 7), rep("factor", 2), rep("ordinal", 2))
test_cor1 <- pairwise_cor(mtcars, type1)
npc_mixed_cor(test_cor1$cor_value, test_cor1$cor_type, test_cor1$cor_p, type1)


}
